ch5/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mlflow-sample-agents"
version = "0.1.0"
description = "MLflowサンプルエージェント - 第5章: 評価"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    # MLflow
    "mlflow>=3.9.0",
    # LangGraph / LangChain
    "langgraph>=1.0.0",
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-community>=0.3.0",
    # ベクトルストア
    "langchain-milvus>=0.2.0",
    "milvus-lite>=2.4.0",
    "setuptools<75",
    # ツール
    "exa-py>=1.0.0",
    "requests>=2.31.0",
    # ユーティリティ
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    # データ取り込み
    "html2text>=2024.0.0",
    "scrapy>=2.11.0",
    "tiktoken>=0.7.0",
    # 評価 (第5章で追加)
    "pandas>=2.0.0",
    "litellm>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agents", "cli", "scripts", "evaluation"]
